+++ unlock-js/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Next Version
 - add `getTokenSymbol` method to web3Service to identify arbitrary ERC20 tokens (#4481)
+- add `getTokenBalance` method to web3Service to get the user's balance of arbitrary ERC20 tokens (#4431)
 
 ## 0.3.10
 - Add "for" field for pending/submitted key purchase transactions (#4190)